Add suggestions: Rename File and Text, Add prefix in Canvas
